+++ nodes/interface.py
@@ -1,5 +1,4 @@
 from copy import copy
-from typing import Any
 import torch
